requirements.txt
Flask
google-cloud-storage
numpy
pandas
Flask-WTF
gensim
nltk